+++ NEWS.md
@@ -2,6 +2,7 @@
 * Correctly parse datetimes that don't include timezone information.
 * Add `icon` and `icon<-` methods for Projects to read the project's current icon URL and to set a new icon by supplying a local file name to upload.
 * Get and set "archived" and "published" status of a dataset with `is.archived`, `is.draft`, and `is.published` (the inverse of `is.draft`). See `?publish` for more.
+* Add `draft` argument to `forkDataset`
 * Support for future API to handle failed long-running jobs.
 
 ## crunch 1.10.0